Backport xrPhysics from X-Ray 1.6

+++ src/xrGame/agent_enemy_manager.cpp
@@ -411,7 +411,7 @@ void CAgentEnemyManager::assign_wounded			()
 {
 	VERIFY					(m_only_wounded_left);
 
-	u32						previous_wounded_count = m_wounded.size();
+	u32						previous_wounded_count = (u32)m_wounded.size();
 	WOUNDED_ENEMY			*previous_wounded = (WOUNDED_ENEMY*)_alloca(previous_wounded_count*sizeof(WOUNDED_ENEMY));
 	std::copy				(m_wounded.begin(),m_wounded.end(),previous_wounded);
 	m_wounded.clear			();
